+++ src/vector.h
@@ -21,19 +21,19 @@ namespace ET
   public:
     Vector();
     ~Vector();
-    Vector(uint64_t dim);
-    Vector(std::string name, uint64_t dim);
+    Vector(uint32_t dim);
+    Vector(std::string name, uint32_t dim);
     Vector(std::vector<T> vec);
     Vector(std::string name, std::vector<T> vec);
-    Vector(uint64_t dim, const T& init);
-    Vector(std::string name, uint64_t dim, const T& init);
+    Vector(uint32_t dim, const T& init);
+    Vector(std::string name, uint32_t dim, const T& init);
 
     //  Getters
-    uint64_t getDim() const;
+    uint32_t getDim() const;
     std::vector<T> getVec() const;
     std::string getName() const;
     //  Setters
-    void setDim(uint64_t dim);
+    void setDim(uint32_t dim);
     void setVec(std::vector<T> vec);
     void setName(std::string name);
 
@@ -62,41 +62,41 @@ namespace ET
     //  the friend method must be defined within the class block.
     friend Vector<T> operator+(T s, const Vector<T>& vector)
     {
-      uint64_t dim = vector.getDim();
+      uint32_t dim = vector.getDim();
       std::string name = "(" + std::to_string(s) + " + "  + vector.getName() + ")";
       Vector<T> v(name,dim,0.0);
-      for (uint64_t i = 0; i < dim; i++) {
+      for (uint32_t i = 0; i < dim; i++) {
           v(i) = vector(i) + s;
       }
       return v;
     }
     friend Vector<T> operator-(T s, const Vector<T>& vector)
     {
-      uint64_t dim = vector.getDim();
+      uint32_t dim = vector.getDim();
       std::string name = "(" + std::to_string(s) + " - "  + vector.getName() + ")";
       Vector<T> v(name,dim,0.0);
-      for (uint64_t i = 0; i < dim; i++) {
+      for (uint32_t i = 0; i < dim; i++) {
           v(i) = s - vector(i);
       }
       return v;
     }
     friend Vector<T> operator*(T s, const Vector<T>& vector)
     {
-      uint64_t dim = vector.getDim();
+      uint32_t dim = vector.getDim();
       std::string name = "(" + std::to_string(s) + " * "  + vector.getName() + ")";
       Vector<T> v(name,dim,0.0);
-      for (uint64_t i = 0; i < dim; i++) {
+      for (uint32_t i = 0; i < dim; i++) {
           v(i) = vector(i) * s;
       }
       return v;
     }
     friend Vector<T> operator/(T s, const Vector<T>& vector)
     {
-      uint64_t dim = vector.getDim();
+      uint32_t dim = vector.getDim();
       std::string name = "(" + std::to_string(s) + " / "  + vector.getName() + ")";
       Vector<T> v(name,dim,0.0);
       std::vector<T> vec(dim);
-      for (uint64_t i = 0; i < dim; i++)
+      for (uint32_t i = 0; i < dim; i++)
       {
         if (vector(i) == 0)
         {
@@ -112,13 +112,13 @@ namespace ET
     }
 
     //  Access operators
-    T& operator()(const uint64_t& i);
-    const T& operator()(const uint64_t& i) const;
+    T& operator()(const uint32_t& i);
+    const T& operator()(const uint32_t& i) const;
 
     std::string summary();
   private:
     //  dimension
-    uint64_t _dim;
+    uint32_t _dim;
     //  container for the coefficients in R^n
     std::vector<T> _vec;
     //  name
@@ -127,7 +127,7 @@ namespace ET
 
   //  Special zero vector
   template<typename T>
-  Vector<T> zeroes(uint64_t dim);
+  Vector<T> zeroes(uint32_t dim);
 
   template class Vector<double>;
 }
